% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{beta_hat}
\alias{beta_hat}
\title{Logistic Regression Coefficients Generation}
\usage{
beta_hat(predictor, response)
}
\arguments{
\item{predictor}{A matrix containing values of the predictors.}

\item{response}{A vector containing values of the responses.}
}
\value{
A vector containing the optimal set of coefficients for logistic regression.
}
\description{
This function generates optimal coefficients for logistic regression.
}
\author{
Cole Kenney
}
