% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{confusion_matrix_and_metrics}
\alias{confusion_matrix_and_metrics}
\title{Confusion Matrix and Metrics}
\usage{
confusion_matrix_and_metrics(predictor, response, cutoff)
}
\arguments{
\item{predictor}{A matrix containing values of the predictors}

\item{response}{A vector containing values of the responses}

\item{cutoff}{A double value - predictions greater than the cutoff are set to one, predictions less than the cutoff are set to zero.}
}
\value{
A list containing the following values:
\describe{
\item{tp}{The double value of true postives. Actual = predicted = 1.}
\item{tn}{The double value of true negatives. Actual = predicted = 0.}
\item{fp}{The double value of false postives. Actual = 0, predicted = 1.}
\item{fn}{The double value of false negatives. Actual = 1, predicted = 0.}
\item{prev}{The double value of prevalence.}
\item{acc}{The double value of accuracy.}
\item{sens}{The double value of sensitivity.}
\item{spec}{The double value of specificity.}
\item{fdr}{The double value of the false discovery rate.}
\item{dor}{The double value of the diagnostic odds ratio.}
}
}
\description{
This function performs logistic regression and then computes a confusion matrix and some common metrics from the confusion matrix.
}
\author{
Cole Kenney
}
