% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{initial_beta_least_squares}
\alias{initial_beta_least_squares}
\title{Initial Coefficients for Logistic Regression}
\usage{
initial_beta_least_squares(X, y)
}
\arguments{
\item{X}{An matrix containing the values of the predictors.}

\item{y}{A vector containing the values of responses.}
}
\value{
A vector containing the initial values of the coefficients for logisitic regression
}
\description{
This function generates the initial coefficients for logistic regression using the least squares formula.
}
\author{
Cole Kenney
}
