% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{plot_metrics_starter_function}
\alias{plot_metrics_starter_function}
\title{Plot (metric)}
\usage{
plot_metrics_starter_function(predictor, response)
}
\arguments{
\item{predictor}{A matrix containing values of the predictors.}

\item{response}{A vector containing values of the responses.}
}
\description{
This function plots (metric) over cutoff values from 0.1 to 0.9.
}
\author{
(name)
}
